[player-4170] -removed withPlayhead and withVideoNavigation HOCs
